requirements.txt
Flask
requests
python-dotenv
Flask-Cors
gunicorn # Gunicorn - это WSGI-сервер, необходимый для продакшена